backend/requirements.txt
fastapi
uvicorn[standard]
python-socketio[asyncio_server]
python-engineio
motor
pydantic
python-dotenv
aiofiles
pandas
